#83 typo with awk

+++ filters/05-delete-long-lines.sh
@@ -41,6 +41,7 @@ candidates=${temp}/files-to-check-line-lengths.txt
 find "${home}" -type f -name '*.java' -print > "${candidates}"
 while read -r f; do
     length=$(LC_ALL=C awk '{ print length($0) }' < "${f}" | sort -n | tail -1)
+    if [ -z "${length}" ]; then continue; fi
     if [ "${length}" -gt "${max}" ]; then
         echo "${f}" >> "${list}"
         rm "${f}"
